An actual (sortof) d3d device on linux!
Major assgoblinry with shared libraries.  There's something fundamental about them I'm just not getting.  They seem like dlls but worse.

Building static causes all sorts of C/C++ problems with linking.  Somehow the Vulkan layer couldn't find the vkGetInstanceProcAddr call which is the main thing the loader does (I think).  I did a nm -D and verified it is in there.  Linked with --verbose and verified that it was linking with the exact file I just checked.  Still undefined reference.

So I ended up calling down into SDL.  It exposes a way to get a pointer to that function, so I just lazy init with that and it works.

Most of the 6 hours it took was fighting with dxvk's meson build system.  I give it a 1 out of 10.  Do we really need to keep making new build systems?  Systems that don't even get the basics right?