Got stalled for a long time on swap chain creation.  The thing that is killing me is that I can't step into SDL from within dxvk.  If I step into SDL from my own code it works.  I'm guessing there's some sort of goblinry going on where the dxvk is loading a different binary.  Why is linux so goofy with shared libs?  I thought dlls couldn't get worse.
Anyway, digging through the assembly I could see some macro failing.  Turns out it was the window handle.  I was grabbing a hwnd through memcpy of some ill defined window struct.  Then I saw a cast of hwnd to SDL_Window... it was one of those freaky C++ casts I never quite trusted, but even magical sky-wizard casting can't do that kind of magic.  I went back and read the docs and yep, you are suppose to just substitute SDL_Window for HWND.

So I have a swapchain!  Yey!

That was a good 6 hours of pain.  I think on windows I'd have had that fixed in 5 minutes, because I'd be able to see what is going on in the debugger.