Kind of starting to get things into library form.
Early on I'll have a testprogram in the root to do simple stuff

Lots of VSCode build / run stuff to get files where they need to go.  I normally don't check in any of those but I'm considering it.

It is very specific to my environment though.  I copy stuff to a game directory with groggy assets to load.

Haven't yet tested if the debugger is able to find all these shared libraries and get the code right.