Compiler tool submodule. Whipped that up last night and it works fairly well.
Beginning to test StuffKeeper

Commiting alot of bogus attempts at shader compilation.  I made a good effort at getting compilation working like it is windows-side where anything that is changed is automagically rebuilt.  Twas not to be though, the compile dll stuff from back in the day isn't open sourced anywhere and so I had to do the external fxc solution.

The submodule uses a fxc in the game's bin folder and then works on stuff in Shaders/ and saves in CompiledShaders/ as usual.

FileStuff is handy file related things

List iterator to return UT_string as well as *

Util to convert to wide strings