forgot #pragma once on some headers
Fix a nasty bug with stripextension and others that directly modify the internal bits of UT_string... didn't realize it was storing the length.  So capping via throwing in a null terminator was making it use the null + the rest of it for keys in the hash table.

A new file in ShaderLib: EntryLayouts.txt.  This associates a vertex shader entry point with one of the types out of Types.hlsli.  Layouts for these are keyed via the type name, so this lets a vertex shader get from the name of the shader to a layout via a couple hash table lookups.

Dictionary has some expanded macro stuff going on while I was trying to debug.  Going to commit it just so I can see how to do it if I need to again, but I'll be removing that.

ccp versions of getvalue (const char pointer not communist)

Create all input layouts that I use (I think) currently.  This was taken care of automagically in all prior versions of groglibs via reflection, but we don't have that here in C land.

Really though the reflection stuff was in the compiler dll which never got open sourced, and the dxvk stuff doesn't take care of it.  Which is why I had to spend a day writing that compile tool.

Will try to make a nice error message in the future for when I make a new shader and forget to create a layout for it.

TestStuff makefile was a bit undercooked.  I've just been mashing F5 to run and hadn't rebuilt in a long time.  Somehow it continued to work despite lots of stuff it depends on changing.

Fought with the LDFLAGS for awhile but they really just don't seem to do anything.  I always just end up getting frustrated and pointing directly at what I want to link.

At least it isn't that meson crap