Lots of StuffKeeper gets that just lazily pull from dictionaries without checking validity
Got rid of most of the random testing stuff in TestStuff.

Spent alot of time trying to get 16 bit floats to work, but it seems it just doesn't and the goog on this is so polluted with bad info I just gave up and made some bit shifts to do it.

Creating a temp vertex buffer with a single triangle so I have something to draw.

Running a timer update loop with SDL event stuff in there.  Not sure I'm doing all that right yet.

Calling draw with a text renderer which probably won't work.  I can see the vert data in renderdoc and the screen is clearing to the right colour though.

Lots more directx calls; State sets, draw, present etc.  Mostly just simple passthrus.

Port of UpdateTimer from C#.  Not entirely sure how functional it is or how accurate.  Will test it more when I have text drawing.