Had a long period of time where I was fiddling and changing anything related to textures, formats, views etc trying to get textures to spit out anything but 0 0 0 0 in the pixel shader.  Eventually it just started working and I have no idea why.
My guess is it might be SamplerDesc MaxLOD 0, or Comparison func never?

Added rotation yaw and pitch to the cube I'm drawing in TestStuff.

Tried having a different format to the render target view of the backbuffer.  I don't think this fixed anything.  I need to have a good think about how to make my swap chain properly.  I want linear color to work in for doing lighting and shadowing and alpha, then a final conversion to SRGB for the screen.  Maybe a postprocess rendertarget in between?

Fix a few more bugs in MakeTexture().  Pals and 32 and 24 all work now.  SetSampler takes a slot.  I'm not sure if this is the S register but I think it is.  Will need this for multiple layers later.