Had some trouble with a basic axial bound ray collide.  Ended up writing it out long form unrolled style so I could debug it better but never quite figured it out.
So I paused and wrote an arbitrary 3D convex shape ray collide and that worked first try.  Axial should be way simpler.  I put in a test where the axial calls the volume and it works, but I'd rather have a working simple axial.

Commiting this as a monument to my stupidity.

Basic ray debug draw code to draw the first cube prim at the collision point.  Cast a ray with the i key.  There's also commented out code for testing bound/plane/volume stuff.  The colour changes to white if no hits, purply if the ray is completely inside the volume, yellowish for an ordinary hit, and cyanish if there's a hit from within the volume.

The plane hit seems to be correct, but I lazily loop and check distances when I should be recording the actual plane used in the intersection.  I'll do that later.  It becomes more important later when doing BSP stuff with plane numbers and the like.

Renamed boxmins/maxs to just mins maxs for the axial bounds of a quadtree node

QuadTree ray intersect is a full recursion to check each leaf right now.  I wanted to fully test and watch the order and make sure the distance checks were good in the debugger and all that.  It seems good.  I'll be adding in bound checks for non leaf nodes as an early out of whole branches of the quadtree.  This makes ray casts touch way less memory.

Frogot to reset num bounds to zero when gathering.  Ended up trying to allocate like a million gigs of ram.

Be sure to reset intersection after checking top root node in QT_LineIntersect

LineIntersectVolume() will return the plane normal hit.

Some return codes from intersections leave the plane in kind of an undefined state.  Like a fully inside doesn't really set the plane anywhere.  Not sure what to do about this yet.

LineIntersectBounds in misc is a giant failure.  I should probably just use a fast reliable proven one from a graphics gems or something.