Invalidate the impact point before re-calling Terrain_LineIntersect.  The function doesn't clear it so you can pass in results from previous tests, like for instance if you have a scenewide raycast with mobiles, bsps involved etc
SSE reciprocal function

A newer ray bound intersect function based on this article I found out on the net:  https://tavianator.com/2011/ray_box.html

I do an additional distance check to make it more line segmenty

Ganked my bad earlier attempt.  The accurate convex volume based intersect is still there and being used for leaf nodes, but really the final shouldn't need a plane returned.  Leaf nodes should just be a hit check and then the ground geometry should be actually intersected and a plane returned for that.  I'll do that soon.

Quadnode intersection now takes a more rayish approach as most of the math will be using the fast intersect.

Need to do a lot of testing to see the difference between the fast / slow and the cost of recomputing an endpoint vs passing it into the recursive function.  I'll probably do a test that shoots 100k random rays or something.  I had something like that back in the C# stuff.

I'm not 100% sure the distance check in the fast raybox is correct.  In my tests flying about shooting rays manually it seemed fine.