Had a bunch of strange situations where I was using mat4 pointers.  Mat4 is a float[][] so already a pointer.  Dunno why I was doing that other than one spot where I was returning one (needs to be a pointer to all that to return it without a warning).
New common prims code in MeshLib.  So far it has a basic xyz axis and an arrow representing a light ray.  The draws still need work.

Renamed some stuff in misc with the Misc_ prefix.

Couple more flexible bounds makers in Misc.  I really need to decide if it is bound or bounds.  Is that like how old people say Walmarts instead of Walmart?

A cube from bounds in PrimFactory.  I made a math mistake when I first wrote this that took a good 4 hours to figure out.  I was looking at the mesh data in renderdoc and was seeing a perfect box.  Turns out I was seeing the bounding box renderdoc draws around stuff.  The actual geometry was a degenerate line from one corner to the opposite.