A new RandomPointInBound() in misc
Wrote some test code to throw 8000 rays at the terrain and time them with rdtsc.  When inside a leaf node I was regenerating the end point of the line segment with a SSE2 reciprocal.  I figured that would be faster than passing in an extra parameter that isn't used unless a leaf is hit, but it turned out to be about 5x faster to pass it in.

Here are the timings on my i5:
Fast node with planar leaf
    Ray test of 8000 rays took 208344141 tics
    Ray test of 8000 rays took 204972273 tics
    Ray test of 8000 rays took 204965919 tics
Fast node with passed in recip
    Ray test of 8000 rays took 58504293 tics
    Ray test of 8000 rays took 49016607 tics
    Ray test of 8000 rays took 48799656 tics
Planar everything
    Ray test of 8000 rays took 800716950 tics
    Ray test of 8000 rays took 790395966 tics
    Ray test of 8000 rays took 791597304 tics

Eventually when the terrain itself is intersected within the leaves, I wont even need the slower plane returning convex volume intersection and all will use the fancy interneto version.  I'll probably use this quick version in my bsp stuff as well when I get back to that.

Next step is to capture the result of many rays and draw them.  I want to go through a large sample visually and make sure there are no corner cases.