a round float to int function in misc
32 bit indexes for terrain.  Needed for anything bigger than 128x128 height texture

Changed the way the quadtree was building.  Instead of passing around TerrainVert I simplified back to a simple grid of heights.  Quadrants are simple integer sizes and such.

With that was able to get rid of BoundVerts(), GetBoundedHeights(), and FixBoxHeights()

Had a nasty bug in the faster raycast.  The reciprocal vector for the ray had the parameter swapped with the endpoint.  That's why it timed so fast in my tests.  It is about the same to recompute the endpoint.

Store leaf collision data as indexed triangles of vec3.  Much smaller.  Haven't written the triangle intersection tests yet.

Rearranged TestStuff a bit, moving all the statics to the end.  Busting some of the keyhandling stuff out into statics.  Will do more cleanup soon.