Record the many rays test and generate one big primitive with the rays as skinny boxes.  Still need to do a set of prims for the impact points.
m key to display the rays

Noticed that a bit of code was using avx instructions.  I hadn't turned it on but decided to try it.  Had some alignment problems at first but fixed those by using aligned_alloc and shuffling the mat4 stuff to the top of the struct.  If __AVX__ is undefined I align on a 16 boundary instead.

I'm probably missing places that will need alignment, and just getting lucky that they happen to land on a 32 byte boundary.

Primitive object destroy.  I'm not cleaning up resources very well thus far.