New crude input library.  Has basic key and mouse bindings and calls a callback with context.
Proper hold down movement and turning for keys

Mouselook with a capture similar to windows setcapture.  SDL takes care of it with relative moves.

Moved anything the input handlers need into a struct that gets passed as context when input is updated

Big gaggle of input handlers at the end of TestStuff.  Usual switch statement is still there but commented out.  I'll get rid of it next commit.