Turn avx back on now that I have that mem stomp bug fixed.  Really it looks like if you have SSE4.2 on avx gets turned on with it.
Move the matrixy portions of Skin to the top of the struct and align the malloc.  I'll probably keep running into alignment stuff as it is a random thing that sometimes alignment happens by chance.

Added a flip parameter to the basic box maker in PrimFactory (for drawing skybox)

Skybox prim and material, follows the camera around.  Doesn't use depth.  Works well with terrain fog.