Rename capsule plane intersect to swept sphere and fix a couple backwards comments
Swept sphere rays now working.  There was a bug in the plane choice where the radius wasn't being subtracted for the distance checks.  This would result in the wrong plane being returned.  Often this would lead to a more distant collision as the returned plane wouldn't have the visible flag