ExpandBoundsByBounds() for the lazy node casts for swept box
Misc_BoundDistanceForNormal(), this tries to find a good amount to shift the plane by to approximate box collision with a single radius

Swept box leaf collision and terrain raycasts.  Works fairly well, not super accurate.  Should be close enough for capsules too.  Box needs to be centered, need to assert for that I guess.