A new biped mover that has a bit of my old C# physics code in it.  It is very goofy right now.  Basic character movement is decent.  Move about and jump and such.  Can still fall through the world on steep surfaces though.
I think the way to fix the fallthrus is to expand the leaf AABB down a
bit.  This should help the swept sphere stuff detect an inside state,
and the hacky pop to surface stuff should fix it.

Hopefully I remember that tomorrow.

The swept sphere move stuff in terrain now returns the "footing".  This
lets you know if you are standing on solid ground, stumbling down
something steep, or midair.