Changed my makefiles again to arch native.  Found out about that from the arch makepkg docs.  Handy!  Of course will have to change it if I release something.  Although the odds of anyone having a cpu more primitive than mine are low.
Don't clamp tiny move vectors unless the character is on the ground.
This was causing the idle animation to kick in when at the peak of the
jump parabola.

Lots of audio init stuff, really just organizing it into create
functions from the main() test stuff I had.  What I didn't realize when
writing this is that FAudio is nothing like the XNA stuff I was porting
this from.  It LOOKS similar but source voices are more like your
hardware channels.  I'll need to rethink all this.

I'll probably just have a fixed set of source voices and when a play
command happens just find a free one and use it.  Then I have to figure
out when the playing is done and free the channel up.