Updated TestStuff.c to work with all the lib changes I've been doing over in tools.
Commented out audio stuff as it is crashing

Commented out halfass post stuff, not sure what I was doing with that.

Flymode is broken

Mainly got this working to see if nappgui was interfering with the game
loop in some way, but it turns out TestStuff has the same problem.

I have this strong feeling of dejavu with this, like I originally had
this problem when I started using SDL and had to change something to fix
it.

I never forked SDL though so I don't know what I would have changed.