Couple asserts on material fetches from library
Trying out an align on a couple stack mat4s (thought that was the
problem)

Check for graphics init fail.  Took me awhile to debug.  SDL2 needs that
vulkan flag set in create window or the swap chain doesn't create
properly.

Bumped refresh down to 60 because I thought it might be causing
problems.  My window was being created on my smaller 60hz monitor.
Should probably be a parameter passed in.