Nuking shapes as it is now merged into UIStuff
Renamed the grogfont functions to GFont_blah

Measure functions in grogfont

Some renames and new vert formats in Layouts.c

When fonts are loaded, attempt to give them a reference to the SRV they
will be using.

UIStuff now handles shapes and fonts together.  Whenever a font changes
the built up verts are drawn.  I'm not super sure this will keep
everything in the correct order vs a ton of individual draw calls but
testing will reveal it.