Color conversion stuff in Misc
Scissor rect stuff in GraphicsDevice

Convert colors passed to the UI stuff to linear

Fiddled with the AlphaBlending state a little bit.  Not sure if I helped
or harmed as the problem I was looking into was color related

Got rid of that annoying vcpkg thing.  I frogot what it was for.