Image drawing
Keep a srv pointer in the UIStuff struct, watch for changes of it or if
text needs to be drawn.  Either way forces a flush of the UI buf.  Seems
to work.

Don't need to pass a font to UI create.  That was from when I had an
idea to have seperate objects for each font.