Added a way to tie a context to an input binding.  Helpful for standalone gumps or something like an inventory bag or such.
Did a macro to use either the update context or the binding context.
This may be the first c macro I've ever written.  It makes me nervous.

The materiallib makefile was still trying to add an include dir for
libpng from when I had it as a submodule

Ran into linker errors when I broke out colladaconvert's skeleton editor
into a seperate file.  It looks like CLAY_IMPLEMENTATION needs to go in
a single c file in the executable being made.  Doesn't really work with
a shared library.