Fiddling with c_cpp_props.  When I added jolt I started getting alot of goofy include errors.  Those are mostly gone but now it doesn't seem to know about true or false.  Oh well.
Keep the heights around in Terrain as a simple float array.

Copy lots of jolt init stuff over from the helloworld sample.

Create a body for the terrain height data.  Not doing anything with it
yet.  Will probably drop spheres with a key and let them roll around a
bit first.