The spheres bouncing around were so easy I was hoping for a miracle of library magic, but this stuff doesn't seem to be ready.
Player movement on the ground seems to be missing bits.  I'm just
getting a segfault and I can't seem to get any debug symbols working so
at this point I'm going to commit what I have and branch.

Think I will try the C++ directly.  Maybe I'll just put a C api to the
stuff I need in PhysicsLib, and compile it with G++ or something.

Wonder if I can rename this branch to joltc