Having more luck with CharacterVirtual.  My feeble brain is very confused by the ++ heavy sample code.  There's a few decent git issues where people go into characters and doing basic movement.  I learned a bit reading those.
Right now the vertical velocity from gravity isn't cleared when the
player lands, so there's a bit of sliding about.  It is tricky because
the player might be standing on a descending platform or somesuch.

I need to have a good look at the samples.

I'm not entirely sure it is doing velocity and gravity correctly.  I
might need some of my own 2 step accumulation stuff.  Jump height seems
really inconsistent.

At least now it is far enough along that I can begin playing around with
the feel of it.  I'm a big fan of the movement in Apex Legends, so
that's kind of the target, even though my current game won't be using
any of that since it is grid based step movement.

The hills too steep to climb bit seems to be working.  The supported
flag is working as well.  I couldn't get that to work with the standard
character. (yet)