Putting the movement physics back into BipedMover.  The jolt stuff just feels spongy and wierd, and I got odd fly-off-into-space stuff that was probably something dumb I was doing.
Anyway, my own stuff feels better.  I think sprint is broken.

Feeding the physics stuff a move vector that is converted to the meters
per second it wants.

After the move, I return a velocity based on the move that happened in
movement-over-frame kind of format.  I was feeding this back into
BipedMover but it was really messing up momentum.  After this commit I'm
going to try an accumulate instead of a direct set.

Moved the stepped updates to 60hz from 144.  The first frame of a drop
at 144hz, gravity has such a tiny effect that it was effectively zero at
144hz, so when I was setting velocity out of the move, I'd gain no
gravitational momentum.  It was super wierd.

The add is sse, and I think a normal fld fadd fst would probably use
more precision internally.

Use the BPM velocity to control animation and the character slerp to
orient to the camera direction.

Before I was using an added up velocity from the stepped updates, but
sometimes, and more-often with 60hz, render updates happen with no
stepped updates, because my framerate is like 3000 or something.

Proper vsync and flip and all that has been on my internal noggin todo
list for a long time.