Playing around with physical constants to test movement stuff
SetVerticalVelocity just copies the Y component for an on ground
situation.

Fixed sprint

Print velocity on screen so I can see what is going on with it

Doing that has revealed that the terrain draw is sort of still there in
a ghostly form when I get that strange missing terrarin problem
sometimes.  Probably a missing or unintialized value in colors or fog or
something like that

After the character move, if still on ground, check the ground plane to
see if it is too steep.  Clear vertical velocity if it is groundy
enough.

Probably also want to change the speed if the groundplane is bad, and
prevent jumping on a bad ground plane like that.

Better animation rate for movement and stopping, but not perfect yet.