This commit is mostly a disaster.  I wanted to keep it as a monument to excessive caffiene.
In Layouts, I'm attempting to support a new graphics format that needs
to match a vertex format to an input layout.  This was alot of work and
went horribly wrong.

I think the best way is just to have a small int array with vert format
elements, and key those off the name of the input layout that uses it
(indirectly).  I was attempting to keep the d3d elements around but they
are not really designed to be more than temporary.  Going to trash and
redo it for the most part.