Constify the match struct a bit for vert searches
Getting rid of most of the C# junk with vert type indexes and such.
Just going to save the name of the vert type.  Still need to remove
those big ass arrays and such.

Mesh_Create() handles the VB and IB and layout stuff I was doing off in
the gltf stuff before.

Keep vert and ind data

Mesh ready to embed within character or static files

MeshPart keeps a direct pointer instead of all that name and dictionary
stuff

Skin create wasn't returning the created skin.  I really wish this was
an error in C and not just a warning.

Statics are handling multiple mesh parts, but still need to pull the
transforms out of wherever they are hiding