I had commented out the localscale read in materials, because I still had a bunch of old matlibs laying around I wanted to read still.  Took me awhile to find that.
Added a way to search for a shader's file.  Like if you give it WNWPos
it should come up with Static

Added a "other way" keyframe to matrix.  I'm guessing these are left vs
right but not sure.  Seems to work well with transforms from blender for
statics.  I suspect the plain GetMatrix is right handed for working with
characters (which I leave right handed until the root node)

Static now created with transforms for each part, need to test moving
the top transform around a bit

Added a float UV flip to half2 in Misc