Found out that bones having different indexes really doesn't seem to happen, so just put in a generic skeleton check.
Decided not to get super complicated with animations of varying
skeletons.  I just warn if a new skelly is more complicated than the
animlib's skel.  The user can just start a new animlib with the new anim
and add all the others and it should work

Proper destroy functions for skelly/animlib/gsnode

Got rid of the dodgy keyframe left hander conversion.  Something is
wrong in all of that stuff because I could never get the fully left hand
path working.  I'm ok with how it is now.

Joint indexes are now bytes

Fix the scale and invscale and scaled root matrices as they are needed
in the non bind bone fetch.  The collision shapes use that to draw and
they were kinda messed up.

Skin save load doing joints array

Fix a bug with the dictionary for when using it to store values instead
of pointers.  It was checking for null as a failure to find something,
but if you have a zero int in there, it coudln't tell the difference the
way I had written it.  Works now!