So I remembered getting sliding working, like if a character falls down a steep slope.  And it was here in this branch.  Not sure why I didn't commit any of this.
I think at the time I needed a sliding animation so I made one, then I
wanted to do a falling anim, and that is the one that I could never get
to export properly with collada.  I think it had to do with the rotation
in the hips?  Anyway that led me to switch over and work on
ColladaConvert (which is now really gltf convert).  And I frogot about
all thsi stuff.

I've since changed stuff alot, so I need to probably merge main back
into this branch, and make sure this stuff still works.

BipedMover splits the movement update into 2 parts.  This allows feeding
back info from the physics system.

Movement code checks the footing to see if player is sliding and animate
that.