Moved all the goofy swapping about and stuff in Anim blend down into the subanims where it is a bit more reasonable.  Gets rid of a const problem as well.
Constify the foreign skeleton in the checker struct

Use int64 for the name to index dictionary.  This is annoying but it
gets rid of warnings.

Warning fixes all over

Removed the const on the vert data passed into Mesh.  I might revisit
this and try to keep it const in the structure because it shouldn't ever
be modified, just read.

Take out the stuffkeeper ref in ParticleBoss so it can be kept const

Logic fix in wrapping menu selection code