Major change to move more towards StructuredBuffer for verts instead of input layouts.
My converter tool is munging verts right now down to a vec4 and uint4.
Alot of places I took out the input layout but never fixed it so a bunch
of stuff is broken, but static meshes and primitives are working.

DrawIndexedInstancedIndirect added to GraphicsDevice

Added a few converters to Misc_ for converting GLTF's normalized 16 bit
colors to floats and such, and a new word interleaver to get 16 bit
half4 values into a format shaders like (so I can use the f16tof32
intrinsic).  I think it uses the old punpckldw instruction underneath.
One of my faves.