Point vert shaders at the right shader in CommonPrims
Get rid of vertex format name in mesh files

Change terrain weights and norms and colours to uint4[4].  ivec is nice
but warnings.

DrawIndexedInstanced for GD.  Not sure I really needed it but maybe this
is the proper way to do a drawcall with no vertex buffer assigned?

Fix the primfactory sphere