All of my makefiles are super messed up.
Everything was always rebuilding everything, so I took the time to fix that in my ludum warmup, but I frogot to do all of them.

Also began on splitting up test programs into smaller bits but didn't get very far.  Ludum is in full swing now so kind of too late.

New c for a ui test program that does nothing so far