Ludum Dare 58 was a disaster.
So the first big problem was the makefile fixes I did right before.  Doing objs as an intermediate step means it can now know when to rebuild stuff and when to skip it.  However this makes it lose the context of what it is building.  Like it doesn't know if the objects are being used in a library or an exe or whatever.  This caused the main function in my audio lib to get built as a real thing.  This caused hours of linker debugging with very vague errors that in no way pointed towards the audio main.  I was so tired by the time I fixed it I have no idea how I even found the problem.  The error messages didn't help.

Next problem is my physics interface has to be a cpp file and I frogot about that and just blindly cuntpasted the same makefile stuff I did to all the others.  This merrily built a nothing file without so much as a warning.  This lead to more vague linker errors.  This one was easier to find as the resulting .so was so small I knew something was up even in my over caffienated state.

The rest of the stuff was art related.  Somehow updating blender a few times over the months completely obliterated any trace that I had ever had GameRig installed.  I had to go back to the previous LD and look at my timelapse video to make sure I wasn't losing my mind.