Newer uthash
Worked a bit more on audio stuff, was kind of a pain to get a list of sounds loaded etc

Had a few locations where I was printing errno but was passing a return value instead of errno

UI test program displays loaded sounds and textures, but needs key and mouse activation

Also would like to hit enter or click to display or play sound effects etc, will do that next

Changed the font ID logic a little bit.  If the id is zero I treat that as wanting to pick a font by size.  I probably should forbid 0 being a natural font ID.