Turn off verbose linker, was using it to debuggify something
UI test can now click textures to display them on the right part of the screen.

Tex list scrolls

I think the font is off in measure or something is off somewhere because the encompassing rounded rects are too big.  Might be using a different font, need to dig into it.