I decided to make it up to the program to set a font ID for drawing clay stuff.  I made a macro to have it also look up the size, as the size is used internally in clay (though not in my stuff).
I might be able to get rid of that and have it automagically fill it in somewhere since my stuff doesn't really scale and a font ID is always a fixed size.

UI test program grabs fonts around size 8, 16, 40 for tiny, medium, big sizes and uses them in the example stuff.

Kinda giving up on the idea of keyboard navigation for this